requirements.txt
streamlit
pillow
requests
pyzbar